fix(rrc): split room history to remove ineffective dynamic import
Move hydrate/clear out of rrcMessagePersist so the session store cycle no longer needs await import().
