+++ vitest.config.ts
@@ -90,6 +90,7 @@ const RENDERER_LOGIC_EXCLUDE = [
   'src/renderer/lib/letsMeshJwt.test.ts',
   'src/renderer/lib/messageRetention.test.ts',
   'src/renderer/lib/rrcMessagePersist.test.ts',
+  'src/renderer/lib/rrcRoomHistory.test.ts',
   'src/renderer/lib/nomad/micronParser.test.ts',
   'src/renderer/lib/nomad/nomadPageCache.test.ts',
   'src/renderer/lib/meshtasticBacklogUtils.test.ts',
